kidney stone detection/kidneydiseasedetection-main/KidneyDisease/streamlit_app/requirements.txt
streamlit
tensorflow
pillow
numpy
opencv-python
pyngrok
